Symmetry: de-randomize torso/neck/head/nose/mouth
I am adding "torso/torso-trans-in|out", "neck/neck-trans-in|out", "head/head-trans-in|out", "nose/nose-trans-in|out" and "mouth/mouth-trans-in|out" so that they're de-randomized following an increase of the Symmetry slider, in the same way as "hip/hip-trans-in|out".
Initially those parameters were partly taken care of by checking for ("trans-horiz" in m.fullName), but then "neck/neck-trans-horiz-in|out", "nose/nose-trans-horiz-in|out" and "torso/torso-trans-horiz-in|out" were renamed in 1.1.0, so the check became ineffective (that's why I am removing it now). It seems that "mouth/mouth-trans-in|out" and "head/head-trans-in|out" weren't handled properly even in 1.0.2.

Also, I am moving the (mGroupName == "head") filter down a bit, so that it doesn't prevent the detection of "head/head-trans-in|out".