changes for a seamless integration of bodyshapes
All targets had been changed to a syntax which can be used by the
default mechanism. Modifier json files are added but must be changed
yet (names of the medical targets look odd). The plugin itself was shortened to standard

descriptions are now shown with mouse over for modifiers.
for filechooser the tooltip mechanism was wrong before, it did not work
correctly. It was meant different before, it should show long asset names in case they did not fit the box.
It never worked because a //2 inside. A text that could be divided into
words was split when the icon was big enough, so theses name had been
visible. Long names without blank or _ were cut and so would be a
problem if they had been different only in the end.

I deleted the tooltip mechanism and add one to display the full label and a line
to show the path. We had too many questions why files had not been shown.
Less information hiding will be helpful, because now the pathname is
visible and the user is able to find the files (if you think it is too much text, it could be configurable maybe)

We still need copy-right statements (better cc0 for targets, otherwise we will again have problems)
But at least the name of Elv should be added to the (atm non-existing) comments inside the  target-files.