images added, license-parser changed
Old makehuman does not save the license created by makeclothes, because
of a ":" at the end of the line. This has been fixed.

Furthermore NO mhxpy had been parsed correctly for a license, because b'author' is not
the same key as "author" ... str(..., "utf-8") was missing.

I added the output of filename, author and license in the material
editor