problems with comments containing non utf-8 codes
A real bad problem happens, when a piece of cloth has a comment in it,
with a non-utf-8 character. The process generating the filecache will
stop and sometimes no or at least less clothes are displayed.

When the asset is selected, the asset is transformed to a mhpxy file and
so the process will also fail!

I fixed that by ignoring errors for the mhclo-file