Enables the 'edit' textbox for all Sliders
Problem: We want to copy/paste a slider's value to another. For instance
we want to copy the left-arm's lowerarm-muscles to the right-arm's
lowerarm-muscles.

Solution: Enables the textbox value editor for all Sliders. Instead of
creating the 'edit' and 'units' widget only when a value converter is
set, we inconditionnaly create them in the constructor. We hide these
widgets in the constructor to not change the current behavior. We add a
'setEditWidgetVisibility()' method so that clients can show/hide the
'edit' widget. We modify the '5_settings_0_settings.py' file to create a
group of radio-button controlling the visibility of the slider's edit
widget. When the visibility is changed we loop through all the slider's
widgets and call the new 'setEditWidgetVisibility()' method. These
radio-buttons control a new setting: 'sliderEditWidgetVisibility'. We
define a default value for this setting in 'mhmain.py'.