Bugfix: combobox is unreadable with 'native' theme
Problem: Combobox is unreadable when using the 'native' theme because
the foreground color is set to white and the 'native' background is
light-grey.

Solution: Remove the hard-coded stylesheet set by the ComboBox
constructor and moves it in the 'makehuman.qss' stylesheet.

Note: Apparently calling setStylesheet in the constructor makes the
stylesheet's value the 'default'. Thus when we later call
'setStylesheet("")' the stylesheet values set by the constructor are
preserved.